+++ packages/builtin-tools/src/index.ts
@@ -9,6 +9,7 @@ import { MemoryManifest } from '@lobechat/builtin-tool-memory';
 import { NotebookManifest } from '@lobechat/builtin-tool-notebook';
 import { PageAgentManifest } from '@lobechat/builtin-tool-page-agent';
 import { SkillsManifest } from '@lobechat/builtin-tool-skills';
+import { LobeToolsManifest } from '@lobechat/builtin-tool-tools';
 import { WebBrowsingManifest } from '@lobechat/builtin-tool-web-browsing';
 import { isDesktop } from '@lobechat/const';
 import { type LobeBuiltinTool } from '@lobechat/types';
@@ -90,6 +91,12 @@ export const builtinTools: LobeBuiltinTool[] = [
   {
     identifier: NotebookManifest.identifier,
     manifest: NotebookManifest,
+    type: 'builtin',
+  },
+  {
+    hidden: true,
+    identifier: LobeToolsManifest.identifier,
+    manifest: LobeToolsManifest,
     type: 'builtin',
   },
 ];
