+++ packages/builtin-tools/src/inspectors.ts
