+++ packages/builtin-tools/src/placeholders.ts
