+++ packages/builtin-tools/src/streamings.ts
