+++ src/features/Conversation/Messages/Tool/Tool/index.tsx
@@ -1,10 +1,10 @@
+import { getBuiltinRender } from '@lobechat/builtin-tools/renders';
 import { Accordion, AccordionItem, Flexbox, Skeleton } from '@lobehub/ui';
 import { type CSSProperties } from 'react';
 import { memo, useState } from 'react';
 
 import Actions from '@/features/Conversation/Messages/AssistantGroup/Tool/Actions';
 import dynamic from '@/libs/next/dynamic';
-import { getBuiltinRender } from '@/tools/renders';
 
 import { dataSelectors, messageStateSelectors, useConversationStore } from '../../../store';
 import Inspectors from '../../AssistantGroup/Tool/Inspector';
