+++ src/server/modules/Mecha/AgentToolsEngine/index.ts
@@ -12,11 +12,10 @@
 import { KnowledgeBaseManifest } from '@lobechat/builtin-tool-knowledge-base';
 import { LocalSystemManifest } from '@lobechat/builtin-tool-local-system';
 import { WebBrowsingManifest } from '@lobechat/builtin-tool-web-browsing';
+import { builtinTools, defaultToolIds } from '@lobechat/builtin-tools';
 import { type LobeToolManifest } from '@lobechat/context-engine';
 import { ToolsEngine } from '@lobechat/context-engine';
 import debug from 'debug';
-
-import { builtinTools, defaultToolIds } from '@/tools';
 
 import {
   type ServerAgentToolsContext,
